requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
tenacity==9.0.0
